requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2 
google-auth==2.29.0
requests==2.31.0